Methods for the 'sdeTMB' R6 class

The following public methods are used to construct a stochastic state space model system, consisting of a set of stochastic differential equations (SDEs), and one or more algebraic observation equations (AOEs). The AOEs are used to infer information about the value of the (latent) states governed by the SDEs, and thus must be functions of at least one state.

Value

The function returns an object of class R6 and sdeTMB, which can be used to define a stochastic state space system.

Methods

Public methods

sdeTMB$new()

sdeTMB$add_systems()

sdeTMB$add_observations()

sdeTMB$add_observation_variances()

sdeTMB$add_inputs()

sdeTMB$add_parameters()

sdeTMB$add_algebraics()

sdeTMB$set_initial_state()

sdeTMB$set_lamperti()

sdeTMB$set_modelname()

sdeTMB$set_cppfile_directory()

sdeTMB$set_map()

sdeTMB$get_systems()

sdeTMB$get_observations()

sdeTMB$get_observation_variances()

sdeTMB$get_algebraics()

sdeTMB$get_initial_state()

sdeTMB$get_parameters()

sdeTMB$construct_nll()

sdeTMB$estimate()

sdeTMB$simulate()

sdeTMB$predict()

sdeTMB$print()

sdeTMB$summary()

sdeTMB$plot()

sdeTMB$clone()

Method new()

Initialize private fields

Usage

sdeTMB$new()

Method add_systems()

Define and add multiple stochastic differential equation governing the process of individual state variables on the form

d<state> ~ f(t,<states>,<inputs>) * dt + g1(t,<states>,<inputs>) * dw1 + g2(t,<states>,<inputs>) * dw2 + ... + gN(t,<states>,<inputs>) * dwN

where f is the drift, and g1, g2, ..., gN are diffusions, with differential brownian motions dw1, dw2, ..., dwN.

Usage

sdeTMB$add_systems(form, ...)

Arguments

form

formula specifying the stochastic differential equation to be added to the system.

...

additional formulas similar to form for specifying multiple equations at once.

Examples

# Specify Ornstein-Uhlenbeck Process
add_systems(dx ~ theta * (mu - x + u) * dt + sigma * dw)
             

Method add_observations()

Define and add a relationship between an observed variable and system states. The observation equation takes the form

<observation> ~ h(t,<states>,<inputs>) + e)

where h is the observation function, and e is normally distributed noise with zero mean and variance to be specified. The observation variable should be present in the data provided when calling estimate(.data) for parameter estimation.

Usage

sdeTMB$add_observations(form, ..., obsnames = NULL)

Arguments

form

formula class specifying the obsevation equation to be added to the system.

...

additional formulas identical to form to specify multiple observation equations at a time.

obsnames

character vector specifying the name of the observation. When the observation left-hand side consists of more than just a single variable name (when its class is 'call' instead of 'name') it will be given a name on the form obs__# where # is a number, unless obsnames is provided.

Examples

#Specify observation directly as a latent state
add_observations(y ~ x)

Specify observation as the sum of exponentials of two latent states
add_observations(y ~ exp(x1) + exp(x2))

Method add_observation_variances()

Specify the variance of an observation equation.

A defined observation variable y in e.g. add_observations(y ~ h(t,<states>,<inputs>) is pertubed by Gaussian noise with zero mean and variance to-be specified using add_observation_variances(y ~ p(t,<states>,<inputs>). We can for instance declare add_observation_variances(y ~ sigma_x^2 where sigma_x is a fixed effect parameter to be declared through add_parameters.

Usage

sdeTMB$add_observation_variances(form, ...)

Arguments

form

formula class specifying the obsevation equation to be added to the system.

...

additional formulas identical to form to specify multiple observation equations at a time.

Method add_inputs()

Declare variables as data inputs

Declare whether a variable contained in system, observation or observation variance equations is an input variable. If e.g. the system equation contains an input variable u then it is declared using add_inputs(u). The input u must be contained in the data.frame .data provided when calling the estimate or predict methods.

Usage

sdeTMB$add_inputs(...)

Arguments

...

variable names that specifies the name of input variables in the defined system.

Method add_parameters()

Declare which variables that are (fixed effects) parameters in the specified model, and specify the initial optimizer value, as well as lower / upper bounds during optimization. There are two ways to declare parameters:

You can declare parameters using formulas i.e. add_parameters( theta = c(1,0,10), mu = c(0,-10,10) ). The first value is the initial value for the optimizer, the second value is the lower optimization bound and the third value is the upper optimization bound.

You can provide a 3-column matrix where rows corresponds to different parameters, and the parameter names are provided as rownames of the matrix. The columns values corresponds to the description in the vector format above.

Usage

sdeTMB$add_parameters(...)

Arguments

...

a named vector or matrix as described above.

Method add_algebraics()

Add algebraic relations.

Algebraic relations is a convenient way to transform parameters in your equations. In the Ornstein-Uhlenbeck process the rate parameter theta is always positive, so estimation in the log-domain is a good idea. Instead of writing exp(theta) directly in the system equation one can transform into the log domain using the algebraic relation add_algebraics(theta ~ exp(logtheta)). All instances of theta is replaced by exp(logtheta) when compiling the C++ function. Note that you must provide values for logtheta now instead of theta when declaring parameters through add_parameters

Usage

sdeTMB$add_algebraics(form, ...)

Arguments

form

formula specifying the stochastic differential equation(s) to be added to the system.

...

additional formulas similar to form for specifying multiple equations at once.

Method set_initial_state()

Declare the initial state values i.e. mean and covariance for the system states.

Usage

sdeTMB$set_initial_state(initial.state, estimate = FALSE)

Arguments

initial.state

a named list of two entries 'x0' and 'p0' containing the initial state and covariance of the state

estimate

boolean value which indicates whether or not the initial conditions shall be estimated as fixed effects parameters. The provided mean and covariance are then used as initial guesses

Method set_lamperti()

Set a Lamperti Transformation

If the provided system equations have state dependent diffusion in a few available ways then it is advantageous to perform a transformation to remove the state dependence. This comes at the cost of a more complicated drift function. The following types of state-dependence is currently supported

'identity' - when the diffusion is state-independent (default)

'log' - when the diffusion is proportional to to x * dw

'logit' - when the diffusion is proportional to x * (1-x) * dw

'sqrt-logit' - when the diffusion is proportional to sqrt(x * (1-x)) * dw

Usage

sdeTMB$set_lamperti(transforms, states = NULL)

Arguments

transforms

character vector - one of either "identity, "log", "logit", "sqrt-logit"

states

a vector of the state names for which the specified transformations should be applied to.

Method set_modelname()

Set modelname used to create the C++ file for TMB

When calling TMB::MakeADFun the (negative log) likelihood function is created in the directory specified by the set_cppfile_directory method with name <modelname>.cpp

Usage

sdeTMB$set_modelname(name)

Arguments

name

string defining the model name.

Method set_cppfile_directory()

Set the path directory where the constructed C++ file is created. You should specify the entire path, unless you want to construct a subfolder in the current working directory - then you can call e.g. set_cppfile_directory("folder_in_current_wd").

Usage

sdeTMB$set_cppfile_directory(directory)

Arguments

directory

string specifying a local directory

Method set_map()

Enable maximum a posterior (MAP) estimation.

Adds a maximum a posterior contribution to the (negative log) likelihood function by evaluating the fixed effects parameters in a multivariate Gaussian with mean and covariance as provided.

Usage

sdeTMB$set_map(mean, cov)

Arguments

mean

mean vector of the Gaussian prior parameter distribution

cov

covariance matrix of the Gaussian prior parameter distribution

Method get_systems()

Retrieve system equations.

Usage

sdeTMB$get_systems()

Method get_observations()

Retrieve observation equations.

Usage

sdeTMB$get_observations()

Method get_observation_variances()

Retrieve observation variances

Usage

sdeTMB$get_observation_variances()

Method get_algebraics()

Retrieve algebraic relations

Usage

sdeTMB$get_algebraics()

Method get_initial_state()

Retrieve initially set state and covariance

Usage

sdeTMB$get_initial_state()

Method get_parameters()

Get initial (and estimated) parameters.

Usage

sdeTMB$get_parameters(type = "all", value = "all")

Arguments

type

one of "all", free" or "fixed" parameters.

value

one of "all", initial", "estimate", "lower" or "upper"

Method construct_nll()

Construct and extract function handlers for the negative log likelihood function.

The handlers from TMB's MakeADFun are constructed and returned. This enables the user to e.g. choose their own optimization algorithm, or just have more control of the optimization workflow.

Usage

sdeTMB$construct_nll(
  data,
  method = "ekf",
  ode.solver = "rk4",
  ode.timestep = diff(data$t),
  loss = "quadratic",
  loss_c = 3,
  unscented_hyperpars = list(alpha = 1, beta = 0, kappa = 3 - private$number.of.states),
  compile = FALSE,
  silent = FALSE
)

Arguments

data

a data.frame containing time-vector 't', observations and inputs. The observations can take NA-values.

method

character vector - one of either "ekf", "ukf" or "tmb". Sets the estimation method. The package has three available methods implemented:

The natural TMB-style formulation where latent states are considered random effects and are integrated out using the Laplace approximation. This method only yields the gradient of the (negative log) likelihood function with respect to the fixed effects for optimization. The method is slower although probably has some precision advantages, and allows for non-Gaussian observation noise (not yet implemented). One-step / K-step residuals are not yet available in the package.

(Continous-Discrete) Extended Kalman Filter where the system dynamics are linearized to handle potential non-linearities. This is computationally the fastest method.

(Continous-Discrete) Unscented Kalman Filter. This is a higher order non-linear Kalman Filter which improves the mean and covariance estimates when the system display high nonlinearity, and circumvents the necessity to compute the jacobian of the drift and observation functions.

All package features are currently available for the kalman filters, while TMB is limited to parameter estimation. In particular, it is straight-forward to obtain k-step-ahead predictions with these methods (use the predict S3 method), and stochastic simulation is also available in the cases where long prediction horizons are sought, where the normality assumption will be inaccurate.

ode.solver

Sets the ODE solver used in the Kalman Filter methods for solving the moment differential equations. The default "euler" is the Forward Euler method, alternatively the classical 4th order Runge Kutta method is available via "rk4".

ode.timestep

the time-step used in the filtering schemes. The time-step has two different uses depending on the chosen method.

Kalman Filters: The time-step is used when numerically solving the moment differential equations.

Laplace Approximation: The time-step is used in the Euler-Maruyama simulation scheme for simulating a sample path of the stochastic differential equation, which serves to link together the latent (random effects) states.

The defined step-size is used to calculate the number of steps between observation time-points as defined by the provided data. If the calculated number of steps is larger than N.01 where N is an integer, then the time-step is reduced such that exactly N+1 steps is taken between observations The step-size is used in the two following ways depending on the chosen method:

Kalman filters: The time-step is used as the step-size in the numerical Forward-Euler scheme to compute the prior state mean and covariance estimate as the final time solution to the first and second order moment differential equations.

TMB method: The time-step is used as the step-size in the Euler-Maruyama scheme for simulating a sample path of the stochastic differential equation, which serves to link together the latent (random effects) states.

loss

character vector. Sets the loss function type (only implemented for the kalman filter methods). The loss function is per default quadratic in the one-step residauls as is natural when the Gaussian (negative log) likelihood is evaluated, but if the tails of the distribution is considered too small i.e. outliers are weighted too much, then one can choose loss functions that accounts for this. The three available types available:

Quadratic loss (quadratic).

Quadratic-Linear (huber)

Quadratic-Constant (tukey)

The cutoff for the Huber and Tukey loss functions are determined from a provided cutoff parameter loss_c. The implementations of these losses are approximations (pseudo-huber and sigmoid approxmation respectively) for smooth derivatives.

loss_c

cutoff value for huber and tukey loss functions. Defaults to c=3

unscented_hyperpars

the three hyper-parameters alpha, beta and kappa defining the unscented transformation.

compile

boolean value. The default (FALSE) is to not compile the C++ objective function but assume it is already compiled and corresponds to the specified model object. It is the user's responsibility to ensure correspondence between the specified model and the precompiled C++ object. If a precompiled C++ object is not found in the specified directory i.e. in <cppfile_directory>/<modelname>/(dll/so) then the compile flag is set to TRUE. If the user makes changes to system equations, observation equations, observation variances, algebraic relations or lamperi transformations then the C++ object should be recompiled.

silent

logical value whether or not to suppress printed messages such as 'Checking Data', 'Building Model', etc. Default behaviour (FALSE) is to print the messages.

Method estimate()

Estimate the fixed effects parameters in the specified model.

Usage

sdeTMB$estimate(
  data,
  method = "ekf",
  ode.solver = "rk4",
  ode.timestep = diff(data$t),
  loss = "quadratic",
  loss_c = 3,
  unscented_hyperpars = list(alpha = 1, beta = 0, kappa = 3 - private$number.of.states),
  control = list(trace = 1, iter.max = 1e+05, eval.max = 1e+05),
  use.hessian = FALSE,
  laplace.residuals = FALSE,
  unconstrained.optim = FALSE,
  compile = FALSE,
  silent = FALSE
)

Arguments

data

data.frame containing time-vector 't', observations and inputs. The observations can take NA-values.

method

character vector - one of either "ekf", "ukf" or "tmb". Sets the estimation method. The package has three available methods implemented:

The natural TMB-style formulation where latent states are considered random effects and are integrated out using the Laplace approximation. This method only yields the gradient of the (negative log) likelihood function with respect to the fixed effects for optimization. The method is slower although probably has some precision advantages, and allows for non-Gaussian observation noise (not yet implemented). One-step / K-step residuals are not yet available in the package.

(Continous-Discrete) Extended Kalman Filter where the system dynamics are linearized to handle potential non-linearities. This is computationally the fastest method.

(Continous-Discrete) Unscented Kalman Filter. This is a higher order non-linear Kalman Filter which improves the mean and covariance estimates when the system display high nonlinearity, and circumvents the necessity to compute the jacobian of the drift and observation functions.

All package features are currently available for the kalman filters, while TMB is limited to parameter estimation. In particular, it is straight-forward to obtain k-step-ahead predictions with these methods (use the predict S3 method), and stochastic simulation is also available in the cases where long prediction horizons are sought, where the normality assumption will be inaccurate.

ode.solver

Sets the ODE solver used in the Kalman Filter methods for solving the moment differential equations. The default "euler" is the Forward Euler method, alternatively the classical 4th order Runge Kutta method is available via "rk4".

ode.timestep

numeric value. Sets the time step-size in numerical filtering schemes. The defined step-size is used to calculate the number of steps between observation time-points as defined by the provided data. If the calculated number of steps is larger than N.01 where N is an integer, then the time-step is reduced such that exactly N+1 steps is taken between observations The step-size is used in the two following ways depending on the chosen method:

Kalman filters: The time-step is used as the step-size in the numerical Forward-Euler scheme to compute the prior state mean and covariance estimate as the final time solution to the first and second order moment differential equations.

TMB method: The time-step is used as the step-size in the Euler-Maruyama scheme for simulating a sample path of the stochastic differential equation, which serves to link together the latent (random effects) states.

loss

character vector. Sets the loss function type (only implemented for the kalman filter methods). The loss function is per default quadratic in the one-step residauls as is natural when the Gaussian (negative log) likelihood is evaluated, but if the tails of the distribution is considered too small i.e. outliers are weighted too much, then one can choose loss functions that accounts for this. The three available types available:

Quadratic loss (quadratic).

Quadratic-Linear (huber)

Quadratic-Constant (tukey)

The cutoff for the Huber and Tukey loss functions are determined from a provided cutoff parameter loss_c. The implementations of these losses are approximations (pseudo-huber and sigmoid approxmation respectively) for smooth derivatives.

loss_c

cutoff value for huber and tukey loss functions. Defaults to c=3

unscented_hyperpars

the three hyper-parameters alpha, beta and kappa defining the unscented transformation.

control

list of control parameters parsed to nlminb as its control argument. See ?stats::nlminb for more information

use.hessian

boolean value. The default (TRUE) causes the optimization algorithm stats::nlminb to use the fixed effects hessian of the (negative log) likelihood when performing the optimization. This feature is only available for the kalman filter methods without any random effects.

laplace.residuals

boolean - whether or not to calculate one-step ahead residuls using the method of oneStepPredict.

unconstrained.optim

boolean value. When TRUE then the optimization is carried out unconstrained i.e. without any of the parameter bounds specified during add_parameters.

compile

boolean value. The default (FALSE) is to not compile the C++ objective function but assume it is already compiled and corresponds to the specified model object. It is the user's responsibility to ensure correspondence between the specified model and the precompiled C++ object. If a precompiled C++ object is not found in the specified directory i.e. in <cppfile_directory>/<modelname>/(dll/so) then the compile flag is set to TRUE. If the user makes changes to system equations, observation equations, observation variances, algebraic relations or lamperi transformations then the C++ object should be recompiled.

silent

logical value whether or not to suppress printed messages such as 'Checking Data', 'Building Model', etc. Default behaviour (FALSE) is to print the messages.

Method simulate()

Perform prediction/filtration to obtain state mean and covariance estimates. The predictions are obtained by solving the moment equations n.ahead steps forward in time when using the current step posterior state estimate as the initial condition.

Usage

sdeTMB$simulate(
  data,
  method = "ekf",
  ode.timestep = diff(data$t),
  ode.solver = "rk4",
  pars = NULL,
  initial.state = self$get_initial_state(),
  n.sims = 100,
  simulation.timestep = diff(data$t),
  k.ahead = 1,
  return.k.ahead = NULL,
  unscented_hyperpars = list(alpha = 1, beta = 0, kappa = 3 - private$number.of.states),
  silent = FALSE
)

Arguments

data

data.frame containing time-vector 't', observations and inputs. The observations can take NA-values.

method

The natural TMB-style formulation where latent states are considered random effects and are integrated out using the Laplace approximation. This method only yields the gradient of the (negative log) likelihood function with respect to the fixed effects for optimization. The method is slower although probably has some precision advantages, and allows for non-Gaussian observation noise (not yet implemented). One-step / K-step residuals are not yet available in the package.

(Continous-Discrete) Extended Kalman Filter where the system dynamics are linearized to handle potential non-linearities. This is computationally the fastest method.

(Continous-Discrete) Unscented Kalman Filter. This is a higher order non-linear Kalman Filter which improves the mean and covariance estimates when the system display high nonlinearity, and circumvents the necessity to compute the jacobian of the drift and observation functions.

All package features are currently available for the kalman filters, while TMB is limited to parameter estimation. In particular, it is straight-forward to obtain k-step-ahead predictions with these methods (use the predict S3 method), and stochastic simulation is also available in the cases where long prediction horizons are sought, where the normality assumption will be inaccurate

ode.timestep

numeric value. Sets the time step-size in numerical filtering schemes. The defined step-size is used to calculate the number of steps between observation time-points as defined by the provided data. If the calculated number of steps is larger than N.01 where N is an integer, then the time-step is reduced such that exactly N+1 steps is taken between observations The step-size is used in the two following ways depending on the chosen method:

Kalman filters: The time-step is used as the step-size in the numerical Forward-Euler scheme to compute the prior state mean and covariance estimate as the final time solution to the first and second order moment differential equations.

TMB method: The time-step is used as the step-size in the Euler-Maruyama scheme for simulating a sample path of the stochastic differential equation, which serves to link together the latent (random effects) states.

ode.solver

Sets the ODE solver used in the Kalman Filter methods for solving the moment differential equations. The default "euler" is the Forward Euler method, alternatively the classical 4th order Runge Kutta method is available via "rk4".

pars

fixed parameter vector parsed to the objective function for prediction/filtration. The default parameter values used are the initial parameters provided through add_parameters, unless the estimate function has been run, then the default values will be those at the found optimum.

initial.state

a named list of two entries 'x0' and 'p0' containing the initial state and covariance of the state

n.sims

number of simulations

simulation.timestep

timestep used in the euler-maruyama scheme

k.ahead

integer specifying the desired number of time-steps (as determined by the provided data time-vector) for which predictions are made (integrating the moment ODEs forward in time without data updates).

return.k.ahead

numeric vector of integers specifying which n.ahead predictions to that should be returned.

unscented_hyperpars

the three hyper-parameters alpha, beta and kappa defining the unscented transformation.

silent

logical value whether or not to suppress printed messages such as 'Checking Data', 'Building Model', etc. Default behaviour (FALSE) is to print the messages.

return.covariance

booelan value to indicate whether the covariance (instead of the correlation) should be returned.

Returns

A data.frame that contains for each time step the posterior state estimate at that time.step (k = 0), and the prior state predictions (k = 1,...,n.ahead). If return.covariance = TRUE then the state covariance/correlation matrix is returned, otherwise only the marginal variances are returned.

Method predict()

Perform prediction/filtration to obtain state mean and covariance estimates. The predictions are obtained by solving the moment equations n.ahead steps forward in time when using the current step posterior state estimate as the initial condition.

Usage

sdeTMB$predict(
  data,
  method = "ekf",
  ode.timestep = diff(data$t),
  ode.solver = "rk4",
  pars = NULL,
  initial.state = self$get_initial_state(),
  k.ahead = 1,
  return.k.ahead = 0:k.ahead,
  return.covariance = TRUE,
  unscented_hyperpars = list(alpha = 1, beta = 0, kappa = 3 - private$number.of.states),
  silent = FALSE
)

Arguments

data

data.frame containing time-vector 't', observations and inputs. The observations can take NA-values.

method

The natural TMB-style formulation where latent states are considered random effects and are integrated out using the Laplace approximation. This method only yields the gradient of the (negative log) likelihood function with respect to the fixed effects for optimization. The method is slower although probably has some precision advantages, and allows for non-Gaussian observation noise (not yet implemented). One-step / K-step residuals are not yet available in the package.

(Continous-Discrete) Extended Kalman Filter where the system dynamics are linearized to handle potential non-linearities. This is computationally the fastest method.

(Continous-Discrete) Unscented Kalman Filter. This is a higher order non-linear Kalman Filter which improves the mean and covariance estimates when the system display high nonlinearity, and circumvents the necessity to compute the jacobian of the drift and observation functions.

All package features are currently available for the kalman filters, while TMB is limited to parameter estimation. In particular, it is straight-forward to obtain k-step-ahead predictions with these methods (use the predict S3 method), and stochastic simulation is also available in the cases where long prediction horizons are sought, where the normality assumption will be inaccurate.

ode.timestep

numeric value. Sets the time step-size in numerical filtering schemes. The defined step-size is used to calculate the number of steps between observation time-points as defined by the provided data. If the calculated number of steps is larger than N.01 where N is an integer, then the time-step is reduced such that exactly N+1 steps is taken between observations The step-size is used in the two following ways depending on the chosen method:

Kalman filters: The time-step is used as the step-size in the numerical Forward-Euler scheme to compute the prior state mean and covariance estimate as the final time solution to the first and second order moment differential equations.

TMB method: The time-step is used as the step-size in the Euler-Maruyama scheme for simulating a sample path of the stochastic differential equation, which serves to link together the latent (random effects) states.

ode.solver

Sets the ODE solver used in the Kalman Filter methods for solving the moment differential equations. The default "euler" is the Forward Euler method, alternatively the classical 4th order Runge Kutta method is available via "rk4".

pars

fixed parameter vector parsed to the objective function for prediction/filtration. The default parameter values used are the initial parameters provided through add_parameters, unless the estimate function has been run, then the default values will be those at the found optimum.

initial.state

a named list of two entries 'x0' and 'p0' containing the initial state and covariance of the state

k.ahead

integer specifying the desired number of time-steps (as determined by the provided data time-vector) for which predictions are made (integrating the moment ODEs forward in time without data updates).

return.k.ahead

numeric vector of integers specifying which n.ahead predictions to that should be returned.

return.covariance

booelan value to indicate whether the covariance (instead of the correlation) should be returned.

unscented_hyperpars

the three hyper-parameters alpha, beta and kappa defining the unscented transformation.

silent

logical value whether or not to suppress printed messages such as 'Checking Data', 'Building Model', etc. Default behaviour (FALSE) is to print the messages.

Returns

A data.frame that contains for each time step the posterior state estimate at that time.step (k = 0), and the prior state predictions (k = 1,...,n.ahead). If return.covariance = TRUE then the state covariance/correlation matrix is returned, otherwise only the marginal variances are returned.

Method print()

Function to print the model object

Usage

sdeTMB$print()

Method summary()

Summary function for fit

Usage

sdeTMB$summary(correlation = FALSE)

Arguments

correlation

boolean value. The default (FALSE) is to not provide the fixed effects parameter correlation matrix.

Method plot()

Function to print the model object

Usage

sdeTMB$plot(
  plot.obs = 1,
  pacf = FALSE,
  extended = FALSE,
  ggtheme = getggplot2theme()
)

Arguments

plot.obs

a vector to indicate which observations should be plotted for. If multiple are chosen a list of plots for each observation is returned.

pacf

logical to indicate whether or not the partial autocorrelations should be returned. The default is FALSE in which case a histogram is returned instead.

extended

logical. if TRUE additional information is printed

ggtheme

ggplot2 theme to use for creating the ggplot.

Method clone()

The objects of this class are cloneable with this method.

Usage

sdeTMB$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples